FIXES
Fixed the CommunityAdapter image loading buug where the Community banner would not load when the app was closed and reopened and the size bug where the banner would be bigger then the default size.

Added permission to read external storage, still didnt do the request methos so the user can accept it.

made the inputType='textPassword' for the login and register layout.